Creación definitiva de la Base de Datos y arreglo de la clase Reparacion (#80)
* Arreglados los atributos duplicados de las clases intermedias y añadidas algunas relaciones al Diagrama de Clases

* Creada la migración de nuevo con todo arreglado

* #79 Arreglado el atributo PrecioTotal de la clase Reparacion (de decimal a float) y creacion de la migración correctamente

+++ src/AppForSEII2526.API/Models/Reparacion.cs
@@ -17,7 +17,7 @@
 
         [Required]
         [DataType(System.ComponentModel.DataAnnotations.DataType.Currency), Display(Name = "Precio Total")]
-        public decimal PrecioTotal { get; set; }
+        public float PrecioTotal { get; set; }
 
         //Relaciones
         [Required]
